+++ wellregistry/registry/admin/monitoring_location.py
@@ -1,11 +1,14 @@
 """
 MonitoringLocation Admin
 """
+import csv
+
+from django.contrib import messages
 from django.contrib.admin import ModelAdmin, RelatedFieldListFilter
 from django.db.models.functions import Upper
 from django.forms import ModelForm, Textarea, ModelChoiceField, HiddenInput
+from django.http import HttpResponse
 from django.urls import path
-
 
 from ..models import MonitoringLocation, AgencyLookup
 from .bulk_upload import BulkUploadView, BulkUploadTemplateView
@@ -56,6 +59,105 @@ class SelectListFilter(RelatedFieldListFilter):
     template = "admin/choice_list_filter.html"
 
 
+CSV_HEADERS = [
+    'Agency',
+    'Site ID',
+    'Site Name',
+    'Latitude',
+    'Longitude',
+    'Horiz Datum',
+    'Horz Loc Method',
+    'Horz Loc Accuracy',
+    'Altitude',
+    'Altitude Units',
+    'Altitude Datum',
+    'Altitude Method',
+    'Altitude Accuracy',
+    'National Aquifer Code',
+    'Local Aquifer Name',
+    'Local Aquifer Cd',
+    'Country',
+    'State',
+    'County',
+    'Well Depth',
+    'Well Depth Units',
+    'Site Type',
+    'Aquifer Type',
+    'Display Well?',
+    'Water Quality Subnetwork',
+    'WQ Baseline Achieved',
+    'WQ Well Characteristics',
+    'WQ Well Type',
+    'WQ Well Purpose',
+    'WQ Purpose Notes (Optional)',
+    'WQ System Name',
+    'Water Level Subnetwork',
+    'WL Baseline Achieved',
+    'WL Well Characteristics',
+    'WL Well Type',
+    'WL Well Purpose',
+    'WL Purpose Notes (Optional)',
+    'WL System Name',
+    'Link'
+]
+
+
+def to_yes_no(flag):
+    """
+    Return true if yes, otherwise no
+    """
+    return 'Yes' if flag else 'No'
+
+
+def get_row(monitoring_location):
+    """
+    Return a list of field values suitable for creating a download csv row.
+    :param monitoring_location:
+    :return: list
+    """
+    return [
+        monitoring_location.agency,
+        monitoring_location.site_no,
+        monitoring_location.site_name,
+        monitoring_location.dec_lat_va,
+        monitoring_location.dec_long_va,
+        monitoring_location.horizontal_datum,
+        monitoring_location.horz_method,
+        monitoring_location.horz_acy,
+        monitoring_location.alt_va,
+        monitoring_location.altitude_units,
+        monitoring_location.altitude_datum,
+        monitoring_location.alt_method,
+        monitoring_location.alt_acy,
+        monitoring_location.nat_aqfr.nat_aqfr_cd if monitoring_location.nat_aqfr else '',
+        monitoring_location.local_aquifer_name,
+        '',
+        monitoring_location.country.country_nm if monitoring_location.country else '',
+        monitoring_location.state.state_nm if monitoring_location.state else '',
+        monitoring_location.county.county_nm if monitoring_location.county else '',
+        monitoring_location.well_depth,
+        monitoring_location.well_depth_units,
+        monitoring_location.site_type,
+        monitoring_location.aqfr_type,
+        to_yes_no(monitoring_location.display_flag),
+        to_yes_no(monitoring_location.qw_sn_flag),
+        to_yes_no(monitoring_location.qw_baseline_flag),
+        monitoring_location.qw_well_chars,
+        monitoring_location.qw_well_type,
+        monitoring_location.qw_well_purpose,
+        monitoring_location.qw_well_purpose_notes,
+        monitoring_location.qw_network_name,
+        to_yes_no(monitoring_location.wl_sn_flag),
+        to_yes_no(monitoring_location.wl_baseline_flag),
+        monitoring_location.wl_well_chars,
+        monitoring_location.wl_well_type,
+        monitoring_location.wl_well_purpose,
+        monitoring_location.wl_well_purpose_notes,
+        monitoring_location.wl_network_name,
+        monitoring_location.link
+    ]
+
+
 class MonitoringLocationAdmin(ModelAdmin):
     """
     Django admin model for the monitoring location
@@ -64,6 +166,8 @@ class MonitoringLocationAdmin(ModelAdmin):
     list_display = ('site_id', 'agency', 'site_no', 'display_flag', 'wl_sn_flag', 'qw_sn_flag',
                     'insert_date', 'update_date')
     list_filter = (('agency', SelectListFilter), 'site_no', 'update_date')
+
+    actions = ['download_monitoring_locations']
 
     @staticmethod
     def site_id(obj):
@@ -96,6 +200,26 @@ class MonitoringLocationAdmin(ModelAdmin):
         """Overrides default implementation"""
         return MonitoringLocation.objects.all() if request.user.is_superuser \
             else MonitoringLocation.objects.filter(agency__in=_get_groups(request.user))
+
+    def download_monitoring_locations(self, request, queryset):
+        """
+        Returns a response which contains a csv with the requested monitoring locations
+        :param request:
+        :param queryset: MonitoringLocation queryset
+        :return: HttpResponse
+        """
+        response = HttpResponse(content_type='text/csv')
+        response['Content-Disposition'] = 'attachment; filename="monitoring_locations.csv"'
+
+        writer = csv.writer(response)
+        writer.writerow(CSV_HEADERS)
+        for monitoring_location in queryset.iterator():
+            writer.writerow(get_row(monitoring_location))
+
+        self.message_user(request, f'Downloaded {queryset.count()} monitoring locations', messages.SUCCESS)
+        return response
+
+    download_monitoring_locations.short_description = 'Download selected monitoring locations'
 
     def has_view_permission(self, request, obj=None):
         """Overrides default implementation"""